Fix comparison between int and float in BeamParticles

+++ game-logic/client/src/main/java/net/bfsr/client/particle/BeamParticles.java
@@ -116,7 +116,7 @@ public class BeamParticles {
                     render.setLastSize();
                 });
 
-        while (lightingParticles.size() < slot.getCurrentBeamRange() / 10.0f) {
+        while (lightingParticles.size() < (int) (slot.getCurrentBeamRange() / 10)) {
             lightingParticles.add(BeamEffects.beamEffect(slot, color));
         }
